+++ examples/orm/app.py
@@ -59,21 +59,21 @@ def get_pets():
 
 
 @app.post('/pets')
-@app.input(PetIn)
+@app.input(PetIn, location='json')
 @app.output(PetOut, status_code=201)
-def create_pet(data):
-    pet = PetModel(**data)
+def create_pet(json_data):
+    pet = PetModel(**json_data)
     db.session.add(pet)
     db.session.commit()
     return pet
 
 
 @app.patch('/pets/<int:pet_id>')
-@app.input(PetIn(partial=True))
+@app.input(PetIn(partial=True), location='json')
 @app.output(PetOut)
-def update_pet(pet_id, data):
+def update_pet(pet_id, json_data):
     pet = PetModel.query.get_or_404(pet_id)
-    for attr, value in data.items():
+    for attr, value in json_data.items():
         setattr(pet, attr, value)
     db.session.commit()
     return pet